requirements.txt
annotated-types==0.7.0
asgiref==3.8.1
certifi==2024.7.4
charset-normalizer==3.3.2
Django==5.1
djangorestframework==3.15.2
gevent==24.2.1
greenlet==3.0.3
idna==3.7
postgres==4.0
psycopg2-binary==2.9.9
psycopg2-pool==1.2
pydantic==2.8.2
pydantic_core==2.20.1
requests==2.32.3
setuptools==72.1.0
SQLAlchemy==0.7.6
sqlparse==0.5.1
supervisor==4.2.5
typing_extensions==4.12.2
urllib3==2.2.2
zope.event==5.0
zope.interface==7.0.1
